migrate ProReportAssignedNotification
